Разделяет reduсer и action на домены
