+++ project/src/components/header/header.tsx
@@ -1,6 +1,6 @@
 import {AppRoute} from 'settings/app-route';
 import {Link} from 'react-router-dom';
-import {resetCityCode} from 'store/action';
+import {resetCityCode} from 'store/city/action';
 import {useAppDispatch} from 'hooks/use-redux-hooks';
 
 function Header(): JSX.Element {
