+++ project/src/components/not-found-screen/not-found-screen.tsx
@@ -1,7 +1,7 @@
 import {AppRoute} from 'settings/app-route';
 import {Header} from 'components/header/header';
 import {Link} from 'react-router-dom';
-import {resetCityCode} from 'store/action';
+import {resetCityCode} from 'store/city/action';
 import {useAppDispatch} from 'hooks/use-redux-hooks';
 
 function NotFoundScreen(): JSX.Element {
