+++ project/src/store/offers/action.ts
@@ -1,0 +1,3 @@
+import {createAction} from '@reduxjs/toolkit';
+
+export const setOffers = createAction('offers/setOffers');
