+++ project/src/store/reducer.ts
@@ -1,32 +1,8 @@
-import {CityCode} from 'types/city-code';
-import {createReducer} from '@reduxjs/toolkit';
-import {DEFAULT_CITY_CODE} from 'settings/const';
-import {Offers} from 'types/offer';
-import {offers as fixtureOffers} from 'fixture/offers';
-import {
-  resetCityCode,
-  setCityCode
-} from 'store/action';
+import {combineReducers} from '@reduxjs/toolkit';
+import {reducer as cityReducer} from 'store/city/reducer';
+import {reducer as offersReducer} from 'store/offers/reducer';
 
-type InitialState = {
-  cityCode: CityCode;
-  offers: Offers;
-}
-
-const initialState: InitialState = {
-  cityCode: DEFAULT_CITY_CODE,
-  offers: fixtureOffers,
-};
-
-const reducer = createReducer(initialState, (builder) => {
-  builder
-    .addCase(resetCityCode, (state) => {
-      state.cityCode = DEFAULT_CITY_CODE;
-    })
-    .addCase(setCityCode, (state, action) => {
-      const {cityCode} = action.payload;
-      state.cityCode = cityCode;
-    });
+export const reducer = combineReducers({
+  cityReducer,
+  offersReducer,
 });
-
-export {reducer};
